Create an AWS role as a secret
Following the DRY principle there should be only one place where an
assume role is specified. Unfortunately, the github terraform provider doesn't
allow creating variables. But there are secrets, so I will use them.

The role secret contains an ARN of a role to be passed in CI, CD
workflows.