Sync web_commit_signoff_required with org level setting
web_commit_signoff_required is enabled on the org level.
Up until now, the github provider applied the org level on a repo if
web_commit_signoff_required isn't set on the repo.
Now it tries to set web_commit_signoff_required to false, which is the
default value on the repo level.

This PR explicitly sets web_commit_signoff_required to true to match
what's already set on the org level.